interview_bot/requirements.txt
gradio>=5.12.0,<6.0
librosa
google-genai==1.10.0
numpy
boto3
python-dotenv
gradio-webrtc
